Updated code structure

+++ handlers/commands/delete.py
@@ -8,10 +8,21 @@ from database.db_manager import DBManager
 db = DBManager()
 
 async def deluser_command(update: Update, context: CallbackContext):
+    chat_type = update.message.chat.type
+    id_del = None
+    id_del = ' '.join(context.args)
+
+    # Ignore messages from groups, supergroups, and channels
+    if chat_type in ["group", "supergroup", "channel"]:
+        return
+    
     user_id = update.message.from_user.id
     if int(user_id) in DEVELOPERS:
-        db.delete_user_data()
-        await update.message.reply_text("✅ All user data has been deleted.")
+        db.delete_user_data(id_del)
+        if id_del:
+            await update.message.reply_text(f"✅ User data has been deleted for this {id_del} id.")
+        else:
+            await update.message.reply_text("✅ All user data has been deleted.")
     else:
         await update.message.reply_text("❌")
         await update.message.reply_text(
@@ -20,6 +31,12 @@ async def deluser_command(update: Update, context: CallbackContext):
         )
 
 async def delfiles_command(update: Update, context: CallbackContext):
+    chat_type = update.message.chat.type
+
+    # Ignore messages from groups, supergroups, and channels
+    if chat_type in ["group", "supergroup", "channel"]:
+        return
+    
     user_id = update.message.from_user.id
     if int(user_id) in DEVELOPERS:
         results = delete_all()
